task 6/todo_car_theme/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
email-validator
Jinja2
python-multipart
psycopg2-binary
